requirements.txt
scikit-learn
fastapi
uvicorn
openai==1.35.14
python-dotenv
PyPDF2
sentence-transformers==2.2.2
moviepy==1.0.3
openai-whisper==20231117
numpy
csv